feat(CU-#3bv0tej) removed hardcoded datas
