+++ src/screens/SearchScreens/TypingSearchScreen/TypingSearchScreen.tsx
@@ -1,14 +1,15 @@
 import React from "react";
 import { Box, Typography } from "@mui/material";
-import { catergoryItemType, SearchResultItemType } from "@forkfacts/models";
-import EggAltOutlinedIcon from "@mui/icons-material/EggAltOutlined";
-import EmojiFoodBeverageOutlinedIcon from "@mui/icons-material/EmojiFoodBeverageOutlined";
-import BookmarkBorderOutlinedIcon from "@mui/icons-material/BookmarkBorderOutlined";
 import { useStyles } from "./typeSearchScreenStyles";
-
+import { GroupListsType, catergoryItemType, SearchResultItemType } from "@forkfacts/models";
 import { SearchResults, RecentSearchHeader, SearchCategories } from "@forkfacts/components";
 
-const TypingSearchScreen: React.FC = () => {
+interface TypingSearchScreenProps {
+  groupLists: GroupListsType[];
+  categoryOptions: catergoryItemType[];
+}
+
+const TypingSearchScreen: React.FC<TypingSearchScreenProps> = ({ groupLists, categoryOptions }) => {
   const handleViewMoreEvent = () => {};
 
   const classes = useStyles();
@@ -45,7 +46,7 @@ const TypingSearchScreen: React.FC = () => {
           categoryOptions={categoryOptions}
         />
         <SearchResults
-          groupLists={GroupLists}
+          groupLists={groupLists}
           grouped={true}
           onSelectItem={onSelectItem}
           onViewMore={handleViewMoreEvent}
@@ -56,35 +57,3 @@ const TypingSearchScreen: React.FC = () => {
 };
 
 export default TypingSearchScreen;
-
-const recentLists = [
-  { image: "/recentImg.png", name: "Kidney beans light, Legume", path: "/:id" },
-  { image: "/image3.png", name: "Grape fruit juices", path: "/:id" },
-  {
-    image: "/image2.png",
-    name: "Baked white bread, Baked products",
-    path: "/:id",
-  },
-  {
-    image: "/image4.png",
-    name: "Grape fruit juice unsweentened, Fruit ...",
-    path: "/:id",
-  },
-  {
-    image: "/image5.png",
-    name: "Banana dehydrated/ banana powder",
-    path: "/:id",
-  },
-];
-
-const GroupLists = [
-  { groupTitle: "FRUIT AND FRUIT JUICES", listItems: recentLists },
-  { groupTitle: "BABY FOODS", listItems: recentLists.slice(0, 1) },
-  { groupTitle: "SWEETS", listItems: recentLists.slice(0, 2) },
-];
-
-const categoryOptions = [
-  { label: "Food", Icon: EggAltOutlinedIcon },
-  { label: "Recipe", Icon: EmojiFoodBeverageOutlinedIcon },
-  { label: "Library", Icon: BookmarkBorderOutlinedIcon },
-];
